Add files via upload
The Dijkstra's algorithm is a well-known graph search algorithm used to find the shortest path between two nodes in a graph. In this case, the metro stations in Delhi were represented as nodes, and the connections between stations were represented as edges in the graph.

To begin the project, I leveraged C++ to build the core logic of the shortest route finder. I implemented the Dijkstra's algorithm in C++, which efficiently calculated the shortest path between the source and destination metro stations. The C++ code was responsible for handling the graph representation, performing the graph traversal, and storing the shortest path information.

Using Python, I developed a command-line or graphical interface where users could input their source and destination metro stations. The Python code interacted with the C++ code to fetch the shortest route information and displayed it to the user.

The integration of C++ and Python allowed for a combination of high-performance computation and an intuitive user interface. C++ provided the optimized implementation of the Dijkstra's algorithm, while Python facilitated user interaction and presentation of the results.